+++ test/Utility/DateCollectionTest.php
@@ -5,6 +5,8 @@ declare(strict_types=1);
 namespace Heptacom\HeptaConnect\Utility\Test;
 
 use Heptacom\HeptaConnect\Utility\Collection\AbstractCollection;
+use Heptacom\HeptaConnect\Utility\Collection\AbstractIterable;
+use Heptacom\HeptaConnect\Utility\Collection\IterableImplementationTrait;
 use Heptacom\HeptaConnect\Utility\Date\Date;
 use Heptacom\HeptaConnect\Utility\Date\DateCollection;
 use PHPUnit\Framework\Attributes\CoversClass;
@@ -12,7 +14,9 @@ use PHPUnit\Framework\Attributes\DataProvider;
 use PHPUnit\Framework\TestCase;
 
 #[CoversClass(AbstractCollection::class)]
+#[CoversClass(AbstractIterable::class)]
 #[CoversClass(DateCollection::class)]
+#[CoversClass(IterableImplementationTrait::class)]
 final class DateCollectionTest extends TestCase
 {
     use ProvidesDateTestsData;