+++ test/Utility/TaggedDateTimeCollectionTest.php
@@ -5,8 +5,10 @@ declare(strict_types=1);
 namespace Heptacom\HeptaConnect\Utility\Test;
 
 use Heptacom\HeptaConnect\Utility\Collection\AbstractCollection;
+use Heptacom\HeptaConnect\Utility\Collection\AbstractIterable;
 use Heptacom\HeptaConnect\Utility\Collection\Contract\AbstractTaggedCollection;
 use Heptacom\HeptaConnect\Utility\Collection\Contract\TagItem;
+use Heptacom\HeptaConnect\Utility\Collection\IterableImplementationTrait;
 use Heptacom\HeptaConnect\Utility\Date\DateTimeCollection;
 use Heptacom\HeptaConnect\Utility\Date\TaggedDateTimeCollection;
 use PHPUnit\Framework\Attributes\CoversClass;
@@ -14,7 +16,9 @@ use PHPUnit\Framework\Attributes\DataProvider;
 use PHPUnit\Framework\TestCase;
 
 #[CoversClass(AbstractCollection::class)]
+#[CoversClass(AbstractIterable::class)]
 #[CoversClass(AbstractTaggedCollection::class)]
+#[CoversClass(IterableImplementationTrait::class)]
 #[CoversClass(TagItem::class)]
 #[CoversClass(DateTimeCollection::class)]
 #[CoversClass(TaggedDateTimeCollection::class)]